refactor(epicenter): reorganize workspace lib structure
Implement Option A (feature-based) from spec 20260203T010500:

Phase 1 - Component moves:
- Move static workspace viewer components from route _components/
  to lib/workspaces/static/components/

Phase 2 - Rename misleading folders:
- Rename lib/docs/ to lib/yjs/ (contains Yjs utilities, not docs)
- Update README.md with correct import paths

Phase 3 - Consolidate query/services/types:
- lib/query/workspaces.ts → lib/workspaces/dynamic/queries.ts
- lib/services/workspaces.ts → lib/workspaces/dynamic/service.ts
- lib/query/static-workspaces.ts → lib/workspaces/static/queries.ts
- lib/services/static-workspaces.ts → lib/workspaces/static/service.ts
- lib/static-workspaces/types.ts → lib/workspaces/static/types.ts
- Add barrel exports for each feature folder

New structure:
lib/
├── workspaces/
│   ├── dynamic/ (queries, service, index)
│   └── static/ (queries, service, types, components/, index)
└── yjs/ (workspace persistence & Yjs utilities)
