+++ apps/epicenter/src/lib/workspaces/dynamic/index.ts
@@ -1,0 +1,2 @@
+export * from './queries';
+export * from './service';
