+++ apps/epicenter/src/lib/workspaces/static/components/StaticWorkspaceSidebar.svelte
