+++ apps/epicenter/src/lib/workspaces/static/components/index.ts
@@ -1,0 +1,3 @@
+export { default as GenericTableViewer } from './GenericTableViewer.svelte';
+export { default as GenericKvViewer } from './GenericKvViewer.svelte';
+export { default as StaticWorkspaceSidebar } from './StaticWorkspaceSidebar.svelte';
