+++ apps/epicenter/src/lib/workspaces/static/index.ts
@@ -1,0 +1,3 @@
+export * from './queries';
+export * from './service';
+export * from './types';
