+++ apps/epicenter/src/lib/workspaces/static/types.ts
